refactor: move text_processor from shared/ to adapters/text/processor.py
- Move text processing service to adapters layer (closer to TextProcessor port)
- Leave backward-compat shim in shared/ for transitive importers
- Update direct importers: audio/recognizer.py, session_service.py
- Add shim allowlist to architecture boundary tests

269 tests pass, lint clean.
